"Fixed date implementation. Rant:"
"INVALID TIME VALUE LINE 361 BECAUSE INDEX6 OF DATE FNS THAT LEADS TO ./ISVALID.JS CHECKS IF THE FORMAT IS A STRING WITH TYPESCRIPT, AND OUR EXAMPLE TASKS ARE NOT NUMBERS BUT STRINGS: WE HAD TO MAKE NEW DATE OBJECTS. JS doesnt process
data in dd/mm/yyyy format - only mm/dd/yyyy,
and not in mm-dd-yyyy either. I was loosing my
god damn sanity working with not knowing what
on earth is the debugger saying and why is
line 361 of date-fns significant.

For anyone reading this, or my future self trying
to figure out how I made this work:
1 - Make the right format mm/dd/yyyy
2 - Parse the string date
3 - Make the date Obj"