"Made a pseudo dom layout for how the dateObjs should work/be displayed"
"Couple of concerns:

1 -How do I go about adding
new subtasks and dates to the objects.. if it's sourced
from multiple cats. The current integration tries
to go the route of - just adding (any new subtasks)
it to whatever task object it came from, which has a
chance of not working as intended.

2 - How to use the least amount of code - without having
to re-invent functions that were already made, and instead
just re-purpose them. Current trend implies that I
might have to do a lot of the implementations again