"Accepted defeat and reality: there is only one way to really make delete task work on a date obj (due to pointers being copied, as we messed up when we parsed the date objs on arrays probably).
That one way is to go to the Task constructor,
and make it trace the cat it belongs to, perhaps
initializing new Task from the Cat, and binding
this.cat = cat.title or something. It will destroy
the last piece of sane and beautiful modular code I
built with task-master, and hence this is not a feasible solution.
"

"Also, working with 2 separate objects, with 1 pointing
towards the other is definately not what I expected, but
God damnit I haven't worked on something that long, that passionately
ever in my entire life. I hope to know the solution one day
to this."